# Widen ascii to utf16