Ready for review now. 

Implements WidenAsciiToUtf16 from #35034

Perf:
```
summary:
better: 1, geomean: 2.172
total diff: 1

No Slower results for the provided threshold = 0.01% and noise filter = 0.3ns.

| Faster                                                       | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Experimental.Perf.IsNormalized_WidenAsciiToUtf16 |      2.17 |         65021.15 |         29935.14 |         |

No file given
```