@kunalspathak 2020-08-10
Same here. Please add the benchmark in `dotnet/performance`.