@kunalspathak 2020-08-10
Could you double check the benchmark `IsNormalized_WidenAsciiToUtf16()` that you added in https://github.com/dotnet/performance/pull/1445 to see if it is actually hitting the code path. You don't do anything with the return value so does JIT optimize away calls that would have otherwise called the code path you wanted to measure?