@pgovind 2020-08-11
Yup, it's hitting the code path. I added an exception that triggered by the benchmarking code:

```
// BeforeAnythingElse


Unhandled exception.
   Cannot print exception string because Exception.ToString() failed.
ExitCode != 0
// Benchmark Process 4827 has exited with code 134
No more Benchmark runs will be launched as NO measurements were obtained from the previous run!
```